+++ app/src/main/res/layout/fragment_guide_first.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+    </data>
+
+    <androidx.appcompat.widget.LinearLayoutCompat
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            style="@style/headline5"
+            android:textColor="@color/secondary_800"
+            android:layout_marginTop="84dp"
+            android:layout_marginHorizontal="20dp"
+            android:text="@string/스팸_문자처럼_쌓여가는_재난_문자로_불편하셨나요_" />
+        <View
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_weight="1"/>
+        <TextView
+            android:id="@+id/btn_next"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            style="@style/button"
+            android:textColor="@color/white"
+            android:layout_marginHorizontal="20dp"
+            android:layout_marginBottom="20dp"
+            android:background="@drawable/bg_radius_4"
+            android:backgroundTint="@color/orange_500"
+            android:paddingVertical="14dp"
+            android:gravity="center"
+            android:text="@string/다음" />
+
+    </androidx.appcompat.widget.LinearLayoutCompat>
+</layout>
