+++ app/src/main/res/layout/fragment_guide_second.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <data>
+
+    </data>
+
+    <androidx.appcompat.widget.LinearLayoutCompat
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <ImageView
+            android:id="@+id/btn_back"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="12dp"
+            app:tint="@color/secondary_300"
+            android:layout_marginLeft="20dp"
+            android:src="@drawable/ic_arrow_left"/>
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            style="@style/headline5"
+            android:textColor="@color/secondary_800"
+            android:layout_marginTop="36dp"
+            android:layout_marginHorizontal="20dp"
+            android:text="@string/스팸_문자처럼_쌓여가는_재난_문자로_불편하셨나요_" />
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_weight="1">
+            <ImageView
+                android:layout_gravity="center"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@drawable/img_logo_on_boarding"/>
+        </FrameLayout>
+        <TextView
+            android:id="@+id/btn_next"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            style="@style/button"
+            android:textColor="@color/white"
+            android:layout_marginHorizontal="20dp"
+            android:layout_marginBottom="20dp"
+            android:background="@drawable/bg_radius_4"
+            android:backgroundTint="@color/orange_500"
+            android:paddingVertical="14dp"
+            android:gravity="center"
+            android:text="@string/다음" />
+
+    </androidx.appcompat.widget.LinearLayoutCompat>
+</layout>
