+++ app/src/main/res/values/strings.xml
@@ -12,6 +12,10 @@
     <string name="후원정보">후원정보</string>
     <string name="후원사_소개">후원사 소개</string>
 
+    <string name="스팸_문자처럼_쌓여가는_재난_문자로_불편하셨나요_">스팸 문자처럼 쌓여가는\n재난 문자로 불편하셨나요?</string>
+    <string name="이제는_대피로에서_알림받고_싶은_재난_유형과_지역을_">이제는 ‘대피로’에서\n알림받고 싶은 재난 유형과 지역을\n맞춤형으로 설정하세요!</string>
+    <string name="다음">다음</string>
+
     <string name="hello_blank_fragment">Hello blank fragment</string>
 
     <string name="네트워크_상태가_불안정합니다_행동요령_및_대피소_조회_화면으로_이동할_수_있습니다_">네트워크 상태가 불안정합니다.\n행동요령 및 대피소 조회 화면으로\n이동할 수 있습니다.</string>
